refactor: usize window indices to drop boundary casts
StockItem and Window held u32 ticker/start indices, forcing u32::try_from on every build from a polars height and as-usize on every tensor slice. polars and burn both index by usize, so switch to usize: removes 7 try_from/expect, 4 as-usize casts, and three # Panics sections. These indices are never serialized, so platform width is a non-issue.

+++ crates/trainer/src/training/batcher.rs
@@ -47,10 +47,8 @@ impl<B: Backend> Batcher<B, StockItem, StockBatch<B>> for StockBatcher<B> {
         let label_slices = items
             .iter()
             .map(|item| {
-                let last = item.start as usize + self.steps - 1;
-                self.labels[item.ticker as usize]
-                    .clone()
-                    .slice(last..last + 1)
+                let last = item.start + self.steps - 1;
+                self.labels[item.ticker].clone().slice(last..last + 1)
             })
             .collect();
         let label = Tensor::cat(label_slices, 0);
